Added registration scenario along with cleaning unnecessary files.
Here i have added the registration controller along with secure password
helper method, so that we can have a password_digest column.
In the application_controller i have added the current_user
helper_method, along with authenticate user method for redirecting user
if he is not involved in the current session.
Also created the user mailer after which user register, a welcome email
will be sent to him from the system.