Added passwords controller through which forgot password scenario is handled.
Although there is too much files here, but mainly i've added the
passwords controller and it's views through which password reset
scenario is handled and authenticated.

I've added the test cases suitable for testing the passwords controller,
the forgot password scenario, reset password mailer & failing scenarios
like empty token and expired token.

Last but not least, i've added the simplecov gem in the gemfile to
calculate the test cases coverage in the codebase.